// frontend/lib/api/models/UserUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserUpdate = {
    full_name?: (string | null);
    bio?: (string | null);
    profile_image?: (string | null);
    preferred_categories?: Array<string> | null;
};